[#41] Rename : redux 모듈 디렉토리 변경
- articles 안에 있던 article, editor 디렉토리를 따로 빼어서 관리하도록 디렉토리 구조를 변경하였습니다.
