+++ frontend/src/modules/articles/articlesSaga.js
@@ -1,4 +1,4 @@
-import createRequestSaga from '../../../services/createRequestSaga.js';
+import createRequestSaga from '../../services/createRequestSaga.js';
 import {
 	DELETE_ARTICLE,
 	READ_ALL_ARTICLES,
@@ -6,7 +6,7 @@ import {
 	READ_MY_ARTICLES,
 	READ_USER_ARTICLES,
 } from './articlesTypes.js';
-import * as articleAPI from '../../../services/api/articleAPI.js';
+import * as articleAPI from '../../services/api/articleAPI.js';
 import { takeLatest } from 'redux-saga/effects';
 
 // 조회
@@ -15,7 +15,7 @@ export const readMyArticlesSaga = createRequestSaga(READ_MY_ARTICLES, articleAPI
 export const readAllArticlesSaga = createRequestSaga(READ_ALL_ARTICLES, articleAPI.getAllArticles);
 export const readUserArticlesSaga = createRequestSaga(
 	READ_USER_ARTICLES,
-	articleAPI.getUserArticles,
+	articleAPI.getMemberArticles,
 );
 // 삭제
 export const deleteArticleSaga = createRequestSaga(DELETE_ARTICLE, articleAPI.deleteArticle);
